Split off GaussianSplatsTool into separate file
