+++ package/Editor/GaussianSplatsTool.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 80d7ecbaa1b24e6399ee95f6fc0b9c90
+timeCreated: 1697718362
